requirements.txt
Django==3.2.10
djangorestframework==3.12.4
drf-yasg==1.20.0